# docs/api.yaml
openapi: 3.0.3
info:
  title: KMS KNPKI
  description: Api Specification for KMS KNPKI
  version: 1.0.0
  contact:
    name: Muhammad Naufal
    email: muhammadnaufaall@gmail.com
    url: https://linkedin.com/in/mhmmdnaufall
servers:
  - url: https://www.knpkid.org/
    description: KNPKI Server
...
paths:
...
  /api/auth/login:
    post:
      summary: Admin login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginAdminRequest"
      responses:
        200:
          description: Success login admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebResponseTokenResponse"
              example:
                data:
                  token: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTcwNzQ3NzEwMCwiZXhwIjoxNzA3NDgwNzAwfQ.I9imemHoNVQiGYAeN8C5_-_1ETTFXOtu_Vg-tAbkkudEi1HiFMqxNOwmt1Ax3ke7rCqiTgh2v96AOaHfRvnzdw
                errors: null
                paging: null

        400:
          description: Wrong Format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebResponse"
              examples:
                blank username:
                  description: username blank error
                  value:
                    data: null
                    errors: "username: must not be blank"
                    paging: null
                blank password:
                  description: password blank error
                  value:
                    data: null
                    errors: "password: size must be between 5 and 30, password: must not be blank"
                    paging: null
                blank username & password:
                  description: username & password blank error
                  value:
                    data: null
                    errors: "password: size must be between 5 and 30, password: must not be blank, username: must not be blank"
                    paging: null
                username length:
                  description: username is not compatible with size format
                  value:
                    data: null
                    errors: "username: size must be between 0 and 20"
                    paging: null
                password length:
                  description: password is not compatible with size format
                  value:
                    data: null
                    errors: "password: size must be between 5 and 30"
                    paging: null
                username & password length:
                  description: username and password is not compatible with size format
                  value:
                    data: null
                    errors: "password: size must be between 5 and 30, username: size must be between 0 and 20"
                    paging: null

        401:
          description: Failed authorizing account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebResponse"
              example:
                data: null
                errors: Username or password wrong
                paging: null
  /api/auth/logout:
    delete:
      summary: Admin logout
      tags:
        - Auth
      security:
        - BearerToken: []
      responses:
        204:
          description: Success logout admin
        401:
          $ref: "#/components/responses/401"
...
components:
  securitySchemes:
    BearerToken:
      description: Put JWT Token with format "Bearer `token`"
      type: apiKey
      in: header
      name: Authorization
  responses:
    401:
      description: User Token Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebResponse"
          example:
            data: null
            errors: Unauthorized
            paging: null
  schemas:
...
    LoginAdminRequest:
      type: object
      properties:
        username:
          type: string
          required: true
          nullable: false
          example: mhmmdnaufal
        password:
          type: string
          required: true
          nullable: false
          example: rahasia
...
    TokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
...
    WebResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
        errors:
          type: string
          nullable: true
        paging:
          $ref: "#/components/schemas/PagingResponse"
    PagingResponse:
      type: object
      nullable: true
      properties:
        currentPage:
          type: integer
          format: int32
        totalPage:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    WebResponseTokenResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TokenResponse'
        errors:
          type: string
        paging:
          $ref: '#/components/schemas/PagingResponse'
...